backend_apps/requirements.txt
tornado==6.0.2
pymongo==3.8.0
redis==3.2.1
neo4j-driver==1.7.4
bleach==3.1.0
pytz==2019.1
Pillow==6.0.0
google-api-python-client==1.7.9
oauth2client==4.1.3
pyDes==2.0.1
boto3